Add desktop, session and revive.  Emacs will now save info when closing and open up in with the same state.  Neato.
Did have to touch ~/.emacs.d/.emacs.desktop though for this to work.  Looking for that file in order to save.

I haven't byte compiled anything yet, and I have to say that the startup with el-get so far is faster than my previous configs.  Less stuff, true, but still pretty good