Oops, missed a commit
Ok, used projectile instead of my projects.  Almost the same thing.  Stopped
the defaults themes from loading, was taking too long.  The slime site appears
gone, so I moved the old slime into site-lisp and configured it.  Seems to be
working, but there is a problem with rainbow-parens.  I'll worry about that
later.