No more multi-term
Why did I use that anyway?  Ansi term works great.  Would like to figure out how to name each buffer accoring to the directory.  There is a function for that, but you have to run it by hand.  Maybe that is good enough for now.

Also some shell tweaks that seem to help even more.